Updater: on macOS, don't uninstall the wake task with the candidate.
The wake task is no longer side-by-side.

Bug: 1339108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4065968
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Auto-Submit: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Xiaoling Bao <xiaolingbao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077645}

+++ chrome/updater/mac/setup/setup.mm
@@ -325,7 +325,6 @@ bool DeleteDataFolder(UpdaterScope scope) {
 void CleanAfterInstallFailure(UpdaterScope scope) {
   // If install fails at any point, attempt to clean the install.
   DeleteCandidateInstallFolder(scope);
-  RemoveUpdateWakeJobFromLaunchd(scope);
   RemoveUpdateServiceInternalJobFromLaunchd(scope);
 }
 
@@ -432,9 +431,6 @@ int UninstallCandidate(UpdaterScope scope) {
     error = kErrorFailedToDeleteFolder;
   }
 
-  if (!RemoveUpdateWakeJobFromLaunchd(scope))
-    error = kErrorFailedToRemoveWakeJobFromLaunchd;
-
   // Removing the Update Internal job has to be the last step because launchd is
   // likely to terminate the current process. Clients should expect the
   // connection to invalidate (possibly with an interruption beforehand) as a
@@ -452,6 +448,9 @@ int Uninstall(UpdaterScope scope) {
 
   if (!RemoveUpdateServiceJobFromLaunchd(scope))
     exit = kErrorFailedToRemoveActiveUpdateServiceJobFromLaunchd;
+
+  if (!RemoveUpdateWakeJobFromLaunchd(scope))
+    exit = kErrorFailedToRemoveWakeJobFromLaunchd;
 
   if (!DeleteInstallFolder(scope))
     exit = kErrorFailedToDeleteFolder;
